Consolidate mongo-join-builder code into adapter-mongoose
