+++ packages/adapter-mongoose/lib/query-parser.js
