+++ packages/adapter-mongoose/lib/tokenizers.js
