--- packages/mongo-join-builder/index.js
@@ -1,4 +1,0 @@
-const { queryParser } = require('./lib/query-parser');
-const { pipelineBuilder } = require('./lib/join-builder');
-
-module.exports = { queryParser, pipelineBuilder };
