--- packages/mongo-join-builder/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "@keystonejs/mongo-join-builder",
-  "description": "Programmatically build JOIN-like Mongo aggregations based on arbitrarily nested relationships.",
-  "version": "8.0.0",
-  "author": "The KeystoneJS Development Team",
-  "license": "MIT",
-  "engines": {
-    "node": ">=10.0.0"
-  },
-  "dependencies": {
-    "@keystonejs/utils": "^5.4.1",
-    "cuid": "^2.1.8"
-  },
-  "devDependencies": {
-    "mongodb": "^3.6.1",
-    "mongodb-memory-server-core": "^6.7.0"
-  },
-  "repository": "https://github.com/keystonejs/keystone/tree/master/packages/mongo-join-builder"
-}
